# obj/Debug/net10.0/EndpointInfo/ClyvoDayApiDocker.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Projeto CLYVO DAY - Documentação OpenAPI",
    "description": "Documentação da API CLYVO DAY - Plataforma de continuidade de cuidado veterinário.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7037"
    },
    {
      "url": "http://localhost:5278"
    }
  ],
  "paths": {
...
    "/api/Clinica/{id}": {
      "get": {
        "tags": [
          "Clinica"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clinica"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/MonitoramentoAnimal": {
      "get": {
        "tags": [
          "MonitoramentoAnimal"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MonitoramentoAnimal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoramentoAnimal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoramentoAnimal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoramentoAnimal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MonitoramentoAnimal/{id}": {
      "get": {
        "tags": [
          "MonitoramentoAnimal"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoramentoAnimal"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoramentoAnimal"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoramentoAnimal"
                }
              }
            }
          }
        }
      }
    },
    "/api/MonitoramentoAnimal/pet/{animalEstimacaoId}": {
      "get": {
        "tags": [
          "MonitoramentoAnimal"
        ],
        "parameters": [
          {
            "name": "animalEstimacaoId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoAnimal"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Tutores": {
      "get": {
        "tags": [
          "Tutores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tutores"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tutores"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tutores"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tutores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Tutores/{id}": {
      "get": {
        "tags": [
          "Tutores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Tutores"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tutores"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tutores"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tutores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Tutores/{id}/atualizacontato": {
      "put": {
        "tags": [
          "Tutores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tutores"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Veterinario": {
      "get": {
        "tags": [
          "Veterinario"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Veterinario"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Veterinario/{id}": {
      "get": {
        "tags": [
          "Veterinario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Veterinario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Veterinario/clinica/{clinicaId}": {
      "get": {
        "tags": [
          "Veterinario"
        ],
        "parameters": [
          {
            "name": "clinicaId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Veterinario/{id}/atualizacontato": {
      "put": {
        "tags": [
          "Veterinario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Veterinario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnimalEstimacao": {
        "required": [
          "name",
          "species",
          "breed",
          "gender",
          "age",
          "weight",
          "birthDate",
          "tutoresId"
        ],
        "type": "object",
        "properties": {
          "animalEstimacaoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "species": {
            "type": "string"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "tutor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Tutores"
              }
            ]
          },
          "tutoresId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "Clinica": {
        "required": [
          "tradeName",
          "cnpj",
          "email",
          "phoneNumber",
          "address"
        ],
        "type": "object",
        "properties": {
          "clinicaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tradeName": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "veterinarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Veterinario"
            }
          }
        }
      },
...
      "MonitoramentoAnimal": {
        "required": [
          "animalEstimacaoId",
          "mood",
          "energyLevel",
          "food",
          "sleepQuality",
          "recentActivities",
          "medication",
          "observations"
        ],
        "type": "object",
        "properties": {
          "monitoramentoAnimalId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animalEstimacaoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animalEstimacao": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AnimalEstimacao"
              }
            ]
          },
          "mood": {
            "type": "string"
          },
          "energyLevel": {
            "type": "string"
          },
          "food": {
            "type": "string"
          },
          "sleepQuality": {
            "type": "string"
          },
          "recentActivities": {
            "type": "string"
          },
          "medication": {
            "type": "boolean"
          },
          "observations": {
            "type": "string"
          },
          "monitoringDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Tutores": {
        "required": [
          "fullName",
          "cpf",
          "email",
          "password",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "tutoresId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "scoreEngagement": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animaisEstimacao": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnimalEstimacao"
            }
          }
        }
      },
      "Veterinario": {
        "required": [
          "fullName",
          "crmv",
          "specialty",
          "email",
          "phoneNumber",
          "clinicaId"
        ],
        "type": "object",
        "properties": {
          "veterinarioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "crmv": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "clinicaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
...
}